Smart contract fixes
