+++ src/Stratis.Bitcoin.Features.SmartContracts/ReflectionExecutor/ReflectionVirtualMachineFeature.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.Extensions.Logging;
+using Stratis.Bitcoin.Builder.Feature;
+using Stratis.Bitcoin.Consensus.Rules;
+
+namespace Stratis.Bitcoin.Features.SmartContracts.ReflectionExecutor
+{
+    public sealed class ReflectionVirtualMachineFeature : FullNodeFeature
+    {
+        private readonly IConsensusRules consensusRules;
+        private readonly ILogger logger;
+
+        public ReflectionVirtualMachineFeature(IConsensusRules consensusRules, ILoggerFactory loggerFactory)
+        {
+            this.consensusRules = consensusRules;
+            this.logger = loggerFactory.CreateLogger(this.GetType().FullName);
+        }
+
+        public override void Initialize()
+        {
+            this.logger.LogInformation("Reflection Virtual Machine Injected.");
+            this.consensusRules.Rules.AddRange(new ReflectionRuleRegistration().GetRules());
+        }
+    }
+}
