+++ src/Stratis.Bitcoin/Consensus/Rules/IConsensusRules.cs
@@ -18,7 +18,7 @@ namespace Stratis.Bitcoin.Consensus.Rules
         /// <summary>
         /// Collection of all the rules that are registered with the engine.
         /// </summary>
-        IEnumerable<IConsensusRule> Rules { get; }
+        List<IConsensusRule> Rules { get; }
 
         /// <summary>
         /// Keeps track of how much time different actions took to execute and how many times they were executed.
